Added Comments to all the Methods

+++ FRC 2013 Robot/src/com/bob85/Climber.java
@@ -19,7 +19,7 @@ public class Climber {
     Joystick rightStick;
     Servo lockClimberServo;
     
-    private double encoderCPR = 250;
+    private int encoderCPR = 250;
     private double encoderDistanceRatio = ((2 * Math.PI) / encoderCPR); //Every encoder revolution is 5.969 linear inches moved on the climber
     
     private double linearClimberMotorOutputCoefficient = -0.058;
@@ -51,7 +51,22 @@ public class Climber {
         rightClimberEncoder.setDistancePerPulse(encoderDistanceRatio);
     }
     
-    
+    /**
+     * Constructs a Climber using: 2 joysticks, 2 encoders, and 4 limitSwitches.
+     * 
+     * @param drive
+     * @param leftStick left Joystick
+     * @param rightStick right Joystick
+     * @param leftClimberMotors inverted driveMotors
+     * @param rightClimberMotors inverted driveMotors
+     * @param leftClimberEncoder left DriveEncoder
+     * @param rightClimberEncoder right DriveEncoder
+     * @param restClimberLimitSwitch switch to show ClimberArm is ready for Shooting
+     * @param extendClimberLimitSwitch switch to show ClimberArm is ready for Climb
+     * @param bottomClimberLimitSwitch bottom HookLimitSwitches
+     * @param topClimberLimitSwitch top HookLimitSwitches
+     * @param lockClimberServo pinLockGear Servo 
+     */
     public Climber(Drive drive, Joystick leftStick, Joystick rightStick,
             Victor leftClimberMotors, Victor rightClimberMotors, 
             Encoder leftClimberEncoder, Encoder rightClimberEncoder,
@@ -73,6 +88,11 @@ public class Climber {
         initEncoderSetting();
     }
     
+    /**
+     * Method used to Shift PTO servos using Buttons 2 & 3 on a Joystick
+     * 
+     * @param joystick Joystick Input 
+     */
     private void joystickBasedShiftClimberLock(Joystick joystick) {
         if (joystick.getRawButton(3)) {
             lockClimberServo.set(1);
@@ -81,6 +101,13 @@ public class Climber {
         }
     }
     
+    /**
+     * Method used to Shift One Side of the PTO
+     * Button 4 shifts leftSide into Climber
+     * Button 5 shifts rightSide into Climber
+     * 
+     * @param joystick 
+     */
     private void joystickBasedShiftOneSide(Joystick joystick) {
         if (joystick.getRawButton(4)) {
             drive.setleftServoClimbPosition();
@@ -91,45 +118,85 @@ public class Climber {
         }
     }
     
+    /**
+     * Get the Boolean from the restingArmSwitch on the climberMount
+     * 
+     * @return 
+     */
     private boolean getIsRest() {
         return restClimberLimitSwitch.get();
     }
     
+    /**
+     * Get the Boolean from the extendedArmSwitch on the climberMount
+     * 
+     * @return 
+     */
     private boolean getIsExtend() {
         return extendClimberLimitSwitch.get();
     }
     
+    /**
+     * Get the Boolean value from the bottomHook limitSwitch
+     * 
+     * @return 
+     */
     private boolean getIsBot() {
         return bottomClimberLimitSwitch.get();
     }
     
+    /**
+     * Get the Boolean value from the topHook limitSwitch
+     * 
+     * @return 
+     */
     private boolean getIsTop() {
         return topClimberLimitSwitch.get();
     }
     
+    /**
+     * Get the Joystick outputs in the climberMode
+     * 
+     * @param joystick 
+     */
     private void getJoystickInput(Joystick joystick) {
         climberMotorOutput = -joystick.getY();
     }
     
+    /**
+     * Stops the Climber Motors
+     */
     private void stopClimb() {
         climberMotorOutput = 0;
         setClimberMotors();
     }
     
+    /**
+     * Changes the Motor output into a linear value correlating to the Joysticks
+     */
     private void setLinearClimbOutput() {
         leftClimberMotors.set(-MotorLinearization.calculateLinearOutput(climberMotorOutput));
         rightClimberMotors.set(MotorLinearization.calculateLinearOutput(climberMotorOutput));
     }
     
+    /**
+     * Sets the Motors to the climberMotorOutput value
+     */
     private void setClimberMotors() {
         leftClimberMotors.set(-climberMotorOutput);
         rightClimberMotors.set(climberMotorOutput);
     }
     
+    /**
+     * Finds the average encoder count value of both Drive Encoders
+     */
     private void calcAvgEncDistance() {
         encoderClimberDistance = ((rightClimberEncoder.getDistance() + leftClimberEncoder.getDistance()) / 2 );
     }
     
+    /**
+     * Computes the average EncoderCounts and resets the values when the bottomLimitSwitch is hit 
+     */
     private void getEncoderDistance() {
         calcAvgEncDistance();
         
@@ -140,6 +207,11 @@ public class Climber {
         }
     }
     
+    /**
+     * Drives the Hooks and sets softLimits on the Hook Movement
+     * 
+     * @param joyStick Joystick input
+     */
     public void manualJoystickElevDrive(Joystick joyStick){
         getEncoderDistance();
         drive.setMotorOutputDeadbands();
@@ -156,6 +228,12 @@ public class Climber {
         }
     }
     
+    /**
+     * Sets the Hooks to drive upwards to an encoder count and stops
+     * 
+     * NOT TESTED, STILL THEORECTICAL
+     * 
+     */
     private void scaleStage1LinearClimberMotorOutputUp() {
         getEncoderDistance();
         
@@ -166,6 +244,11 @@ public class Climber {
         }
     }
     
+    /**
+     * Drives the ClimberHooks moves downward in a scaling fashion
+     * 
+     * NOT TESTED, STILL THEORETICAL 
+     */
     private void scaleStage2LinearClimberMotorOutputDown() {
 
         if (!bottomClimberLimitSwitch.get()) {
@@ -180,6 +263,13 @@ public class Climber {
         }
     }
     
+    /**
+     * Sets a Button to a certain height in encoderCounts
+     * 
+     * @param joystick Joystick Input
+     * @param topButton Button to be pressed
+     * @param presetHeight an Encoder Value set as a Max Height Setpoint
+     */
     public void setClimberToPresetHeight(Joystick joystick, int topButton, double presetHeight){
         if (joystick.getRawButton(topButton) == true){
             if (presetHeight > encoderClimberDistance && topClimberLimitSwitch.get() != true){
@@ -191,8 +281,13 @@ public class Climber {
         }
     }
     
-    
-
+    /**
+     * Case 0 = sets the Hooks on a Manual Control
+     * Case 1 = auto drives up
+     * Case 2 = auto drives down
+     * 
+     * NOT TESTED, STILL THEORECTICAL 
+     */
      public void runAutoClimb() {
         
         switch(climberStage) {
@@ -231,7 +326,10 @@ public class Climber {
                  stopClimb();
          }
     }
-     
+    
+    /**
+     * Sets the parameters for switching into Manual Climb
+     */
     public void switchClimbStates() {
         if (drive.getDriveState() == 3) {
             climberState = 1;
@@ -240,6 +338,11 @@ public class Climber {
         }
     }
     
+    /**
+     * Case 0 = do nothing
+     * Case 1 = set the Joysticks for manualHookDriving
+     * 
+     */
     public void runClimbStates() {
         switch (climberState) {
             case 0:
@@ -251,17 +354,26 @@ public class Climber {
         }
     }
      
+    /**
+     * Puts values for LimitSwitches on the SmartDashboard
+     */
     public void runDiagnostics() {
         SmartDashboard.putBoolean("Rest LimitSwitch", getIsRest());
         SmartDashboard.putBoolean("Extend LimitSwitch", getIsExtend());
         SmartDashboard.putBoolean("Bot LimitSwitch", getIsBot());
         SmartDashboard.putBoolean("Top LimitSwitch", getIsTop());
     } 
-     
+    
+    /**
+     * Set values when starting the Robot
+     */
     public void initClimber() {
         
     } 
-     
+    
+    /**
+     * Executes Methods for the Main Robot Class
+     */
     public void runClimber() {
         runDiagnostics();
         switchClimbStates();
